pyproject.toml
[tool.poetry]
name = "vuanem_sheets_pg_boilerplate"
version = "0.1.0"
description = ""
authors = ["HM <hieumdd@gmail.com>"]

[tool.poetry.dependencies]
python = "~3.9"
google-api-python-client = "^2.27.0"
google-auth-httplib2 = "^0.1.0"
google-auth-oauthlib = "^0.4.6"

[tool.poetry.dev-dependencies]
black = "^21.9b0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
